add vs2005 project files for backward compatility
git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1885 f17a810d-92cc-4fca-aeb8-6e32867107aa

+++ itpp/base/random_dsfmt.h
@@ -171,7 +171,7 @@ public:
       dsfmt_gen_rand_all();
       idx = 0;
     }
-    return psfmt64[idx++] & 0xffffffffU;
+    return (uint32_t)(psfmt64[idx++] & 0xffffffffU);
   }
 
   /*!
